media: video: Make VCAP properly call regulator_get
VCAP was not providing a pointer to the device when calling
regulator_get. Although the call was able to get the regulator,
there was always an error message regarding not being able to
create a debugfs node. The problem was because the device provided
to the regulator_get call was NULL.

Change-Id: I6b344cafd96db50e95abfbf46021a361ac0de638
Signed-off-by: Terence Hampson <thampson@codeaurora.org>