media: video: Prevent buffer ion resource from being freed twice
It was possible for the client to have two threads running where
one would be blocked in dq_buf that would wake up while the other
thread was in the process of turning the stream off. This cause
two thread to try freeing the ion resource which would result
in a crash.

Change-Id: Ibbbf34c7770eed99e1a89d2a4c41ad48f5b6a13a
Signed-off-by: Terence Hampson <thampson@codeaurora.org>