arm/dt: msm8974: Support system wakeup via SDCC DAT1 line
All SDCC slots in 8974 support system wakeup capability
when the SDIO card asserts the DAT1 line low during system
sleep. To support this in the driver add DT properties
specifying the interrupt number.

For SDC1 and SDC2 slots the DAT1 line is a dedicated MSM
pin so use "qcom,dat1-mpm-int" to specify MPM interrupt
line mapping.

For SDC3 and SDC4 slots the DAT1 line is an MSM GPIO, so
use "interrupt-map" and corresponding  properties to specify
the gpio to irq mapping.

Change-Id: I7ebac4ed4f4e7d325443c4a29d6dc4089f333c56
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>