gpio: msm: Get rid of NR_GPIO_IRQS
Remove the use and definition of the NR_GPIO_IRQS macro for
device tree targets. This involves removing the macro from the
gpio driver as well. Since this a common gpio driver, certain
non-device tree targets had to be fixed to support this.
However, the NR_GPIO_IRQS macro continues to exist for the
non-device tree targets.

Change-Id: Ie05b20e9f7732f3eb9c972d2aa81280f4b736f29
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>