msm: pil-q6v5-mss: Add support for non-loadable modems
The MSM9625 has the same modem as the MSM8974. However, the MSM9625
modem is not loaded by PIL, but reports errors in the same way as
before. Therefore, introduce a device tree parameter to specify
whether the modem is loadable, and only initialize the PIL device
when this parameter is set.

Separate out PIL and subsystem initialization to individual
functions to make this code cleaner.

Also, to ensure that the MSM8974 modem still boots, add the parameter
to the modem device tree node for the MSM8974.

Change-Id: Iab2a1e8a661f160d492bfc96e5e19433ae0199e4
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>