msm: clock: Allow zero rates on 8974 and 9625 UART clocks
In certain chip configurations, sharing of the UART instance
between processors results in the UART clock registers being
programmed without the knowledge of the Linux clock driver.

Thus it is required that the clock rate be reprogrammed. To
allow this, add in a zero rate to the UART clock frequency
tables, so that setting the rate to zero and then back to
the original rate will cause the Linux clock driver to
actually reprogram the registers.

CRs-Fixed: 418578
Change-Id: I99986a827215380e90da97df11437d870c627830
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>