hwmon: qpnp-adc: Add Reverse calibration
On VADC_TM peripheral, clients have the ability
to set thresholds and receive interrupts when the
thresholds are crossed. These thresholds are required
to be converted to voltages if its a temperature
threshold and calibration is to be applied on them.
Depending on the channels the voltage and gain
api returns the absolute or ratiometric calibration
values for scaling functions. The scaling functions
currently supported are for channels usb_id, batt_therm
and thermistor channels using 100k pull-ups.

Change-Id: I1973f5b3fa81a3e668baed253a501b5ab3256e98
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>