arm/dt: msm8974: Fix throughput requirements for SDCC bus scaling
SDCC throughput requirements in 8974 for bus scaling are
represented in Kilo Bytes per second but the bus scaling
driver actually expects them to be in Kilo Bits per second
and converts into Bytes per second according to bus width
of the buses in the voting path. Fix this by representing
the throughput values in kilo bits per sec. Also, update
the throughput requirements according to clock plan and
not w.r.t mode in which the SD/eMMC card operates.

Change-Id: I32ae34a60159bbdd013b7111166993f8a093a9bf
Signed-off-by: Sujit Reddy Thumma <sthumma@codeaurora.org>