mmc: sdio: fix resume failure
If SDIO keep power flag (MMC_PM_KEEP_POWER) is not set, card would
be reinitialized during resume but as we are not resetting
(CMD52 reset) the SDIO card during this reinitialization, card may
fail to respond back to subsequent commands (CMD5 etc...).

This change resets the card before the reinitialization of card
during resume.

CRs-Fixed: 424805
Change-Id: Ic4bf6fe51dbfe8722ed7bbd9f98d7a650ed6e5d0
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>