mmc: msm_sdcc: don't print suspend/resume statistics in case of errors
Currently we print the time taken during SDCC host driver suspend/resume
(both platform and runtime) even if we encountered errors during it.
But sometimes, for example runtime suspend might be aborted because
actual SDCC transfer request is arrived at the same time. So printing
the statistics in such cases will not be helpful and may even cause the
confusion.

This change prints the time taken in suspend/resume only when those
operations are successfully completed.

Change-Id: I7fa5754ccddee0d90f684b05526e9434c068a0ff
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>