msm: ramdump: Dump ELF format instead of plain memory
Ramdump devices dump memory directly, relying on the users of
the device nodes to know that the file they read may represent
non-contiguous memory segments. In addition, they must know what
addresses those segments pertain to, something that is not easily
determined with relocatable PIL.

Introduce a new API, do_elf_ramdump(), that generates an ELF core
file when you read the device node. This should allow users of
the ramdump device nodes to easily figure out what the ramdump
represents instead of relying on hard-coded knowledge. It also
allows one to use readelf on the file to determine what the
memory regions are, and simplifies loading of the binaries into
debuggers that can understand ELF formats.

Change-Id: I530f72298c7e88fe95c020c6a9b99d8160df2f4c
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>