msm: vidc: Make wait_for_completion uninterruptable
If userspace process dies when driver is waiting for response from the
core, the wait gets interrupted.  This causes race conditions between
the HLOS and the Venus core, especially in cases of buffer frees.
Hence, switch to an uninterruptable wait so that userspace can't cause
the driver to misbehave.

Also reduce the hardware timeout to 200 ms, since being uninterruptable
for 5 minutes is a bad idea.

Change-Id: If51a700a0b64cb240bce9e12755a28b6acd80279
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>