power: pm8921-bms: Relax UUC delta restriction based on time
Previously, the Unusuable Charge calculation was restricted
to changing by 1% every time it was recalculated in order to
meet linearity demands.

However, if the phone is asleep, it might take much longer
than 20 seconds for the UUC to be recalculated, in which it
could have drifted a lot more than 1% due to temperature
changes.

Fix this by allowing UUC to change by an additional 1% for
every 60 real seconds between successive UUC calculations.

Change-Id: Ic1e7cf4da4d2bf73d889aee7bac6cf260b26d532
Signed-off-by: Xiaozhe Shi <xiaozhes@codeaurora.org>