msm: ipa: fix data stall caused by IPA HW bottleneck.
When a receiver is not able to keep up with
the load pumped by the IPA SW bridge,
the desc FIFO gets full and subsequent SPS
transfers fail.
Unless bridge polls for completed descriptors
in this state, data path will remain stuck indefinitely.
This commit adds support for that polling and
symbol renaming (symbols such as functions
names and variables).

Change-Id: Iddb187798403a8c91e66f4c2d515fc54f983cf67
Signed-off-by: Talel Atias <tatias@codeaurora.org>